% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTION_calculate_cone_metrics.R
\name{calculate_cone_metrics}
\alias{calculate_cone_metrics}
\title{Function to calculate both volume and surface area of a cone with radius variations}
\usage{
calculate_cone_metrics(h_i)
}
\arguments{
\item{h_i}{Numeric. The total height of the cone.}
}
\value{
A list with two elements:
  \describe{
    \item{volume}{Numeric. The computed volume of the cone.}
    \item{surface_area}{Numeric. The computed surface area of the cone.}
  }
}
\description{
This function calculates the volume and surface area of a cone, where the radius varies along the height due to random perturbations. The calculation uses numerical integration.
}
\examples{
calculate_cone_metrics(10)
calculate_cone_metrics(h_i = 10)
calculate_cone_metrics(5)
calculate_cone_metrics(h_i = 5)
}
\author{
Lodrik Adam
}
