% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APP_ice_cream_party.R
\name{cityModuleServer}
\alias{cityModuleServer}
\title{City Module Server Function}
\usage{
cityModuleServer(id, api_key_reactive)
}
\arguments{
\item{id}{A unique identifier matching the module's UI `id`.}

\item{api_key_reactive}{A reactive expression containing the API key for accessing the OpenWeatherMap API.}
}
\value{
Server logic for the city module.
}
\description{
Server logic for the city module, handling data retrieval from the OpenWeatherMap API, running simulations, and generating outputs.
}
\author{
Lodrik Adam
}
\keyword{internal}
