% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APP_ice_cream_party.R
\name{cityModuleUI}
\alias{cityModuleUI}
\title{City Module UI Function}
\usage{
cityModuleUI(id)
}
\arguments{
\item{id}{A unique identifier for the module's namespace.}
}
\value{
A `tabPanel` containing the UI elements for the city module.
}
\description{
Creates the UI for the city module, allowing users to input a city name, select a date, choose parameters, and view results.
}
\author{
Lodrik Adam
}
\keyword{internal}
