% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTION_cone_radius_for.R
\name{cone_radius_for}
\alias{cone_radius_for}
\title{Cone Radius Calculation (Vectorized using for loop)}
\usage{
cone_radius_for(x_vals)
}
\arguments{
\item{x_vals}{A numeric vector of input height values for which the cone radius will be calculated.}
}
\value{
A numeric vector containing the calculated cone radius corresponding to each input height values in `x_vals`.
}
\description{
This function calculates the radius of a cone for each height value in a numeric vector `x_vals`. It applies conditional logic to determine the radius based on the input value.
}
\examples{
cone_radius_for(c(-1, 4, 8, 8 + pi / 4, 10))
cone_radius_for(c(-5, 15))
cone_radius_for(seq(0, 10, by = 2))
}
\author{
Lodrik Adam
}
