% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTION_cone_radius_map.R
\name{cone_radius_map}
\alias{cone_radius_map}
\title{Cone Radius Calculation (Vectorized using map)}
\usage{
cone_radius_map(x_vals)
}
\arguments{
\item{x_vals}{A numeric vector of input height values for which the cone radius will be calculated.}
}
\value{
A numeric vector containing the calculated cone radius corresponding to each input height value in `x_vals`.
}
\description{
This function calculates the radius of a cone for each height value in a numeric vector `x_vals`. It uses the `map_dbl` function from the `purrr` package.
}
\examples{
cone_radius_map(c(-1, 4, 8, 8 + pi / 4, 10))
cone_radius_map(c(-5, 15))
cone_radius_map(seq(0, 10, by = 2))
}
\author{
Lodrik Adam
}
