% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTION_cone_radius_non_vectorized.R
\name{cone_radius_non_vectorized}
\alias{cone_radius_non_vectorized}
\title{Cone Radius Calculation (Non-Vectorized)}
\usage{
cone_radius_non_vectorized(x)
}
\arguments{
\item{x}{A numeric value representing the input height to calculate the cone radius. Must be a single number.}
}
\value{
A numeric value representing the calculated radius for the given height. Returns 0 for inputs outside the defined range.
}
\description{
This function calculates the radius of a cone based on the input height `x` according to specific conditions.
}
\examples{
cone_radius_non_vectorized(4)
cone_radius_non_vectorized(9)
cone_radius_non_vectorized(15)
}
\author{
Lodrik Adam
}
