% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTION_cone_radius_sapply.R
\name{cone_radius_sapply}
\alias{cone_radius_sapply}
\title{Cone Radius Calculation (Vectorized using sapply)}
\usage{
cone_radius_sapply(x_vals)
}
\arguments{
\item{x_vals}{A numeric vector of input height values for which the cone radius will be calculated.}
}
\value{
A numeric vector containing the calculated cone radius corresponding to each input height value in `x_vals`.
}
\description{
This function calculates the radius of a cone for each height value in a numeric vector `x_vals` using `sapply`. It applies conditional logic to determine the radius for each input value.
}
\examples{
cone_radius_sapply(c(-1, 4, 8, 8 + pi / 4, 10))
cone_radius_sapply(c(-5, 15))
cone_radius_sapply(seq(0, 10, by = 2))
}
\author{
Lodrik Adam
}
