% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTION_cone_radius_vectorized.R
\name{cone_radius_vectorized}
\alias{cone_radius_vectorized}
\title{Cone Radius Calculation (Vectorized version using Vectorize function)}
\usage{
cone_radius_vectorized(x)
}
\arguments{
\item{x}{A numeric height value representing the input for which the cone radius will be calculated.}
}
\value{
A numeric value representing the calculated radius for the height input. Returns 0 for inputs outside the defined range.
}
\description{
This function calculates the radius of a cone based on a single input height value `x` using vectorized logic. The function handles specific conditions to determine the radius based on the input value. It uses the `cone_radius_non_vectorized`function.
}
\examples{
cone_radius_vectorized(4)
cone_radius_vectorized(8 + pi / 4)
cone_radius_vectorized(-1)
cone_radius_vectorized(10)
}
\author{
Lodrik Adam
}
