% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APP_ice_cream_party.R
\name{run_ice_cream_party_planner}
\alias{run_ice_cream_party_planner}
\title{Launch the Ice Cream Party Planner Shiny Application}
\usage{
run_ice_cream_party_planner()
}
\value{
Lauches the Shiny application for analyzing climate factors and simulating ice cream cone requirements. The app provides the following features:
  \itemize{
    \item API Key Input: Securely enter and store your OpenWeatherMap API key.
    \item City Modules: Three separate modules to analyze different cities simultaneously.
    \item Weather Forecast: View 5-day forecasts for temperature, humidity, and pressure.
    \item Simulations: Run simulations to determine ice cream cone requirements based on the selected weather metrics.
    \item Visualizations: Interactive plots and histograms to visualize forecast data and simulation outcomes.
  }
}
\description{
This function initializes and runs the Shiny application for analyzing climate factors
and simulating ice cream cone requirements based on weather forecasts. Upon launching,
users are prompted to enter their OpenWeatherMap API key. The application allows users
to input a city name, select a date, choose weather parameters to display, and view
forecast data alongside simulation results for ice cream volume and surface area for up
to 3 cities simultaneously.

Ensure you have a valid OpenWeatherMap API key before launching the application.
The application utilizes parallel processing to enhance simulation performance.
}
\examples{
\dontrun{
run_ice_cream_party_planner()
}
}
\author{
Lodrik Adam
}
