% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APP_ice_cream_party.R
\name{server}
\alias{server}
\title{Shiny Application Server Logic}
\usage{
server(input, output, session)
}
\arguments{
\item{input}{The input values from the Shiny application.}

\item{output}{The output values from the Shiny application.}

\item{session}{The Shiny session object.}
}
\value{
Server logic for the Shiny application.
}
\description{
Defines the server logic for the Shiny application.
}
\author{
Lodrik Adam
}
\keyword{internal}
