% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APP_ice_cream_party.R
\docType{data}
\name{ui}
\alias{ui}
\title{Shiny Application UI}
\format{
An object of class \code{shiny.tag.list} (inherits from \code{list}) of length 4.
}
\usage{
ui
}
\value{
A `fluidPage` containing the application's UI.
}
\description{
Defines the user interface for the Shiny application.
}
\author{
Lodrik Adam
}
\keyword{internal}
